app: create flow helpers, implement Flow.waitFor
